+++ R/tar_map2_count_raw.R
@@ -19,7 +19,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
-#' @inheritSection tar_rep Nested futures for batched replication
+#' @inheritParams tar_rep
 #' @param batches Positive integer of length 1,
 #'   maximum number of batches (dynamic branches within static branches)
 #'   of the downstream (`command2`) targets. Batches
@@ -62,6 +62,7 @@ tar_map2_count_raw <- function(
   combine = TRUE,
   suffix1 = "1",
   suffix2 = "2",
+  parallel_reps = FALSE,
   columns1 = quote(tidyselect::everything()),
   columns2 = quote(tidyselect::everything()),
   tidy_eval = targets::tar_option_get("tidy_eval"),
@@ -98,6 +99,7 @@ tar_map2_count_raw <- function(
     suffix2 = suffix2,
     columns1 = columns1,
     columns2 = columns2,
+    parallel_reps = parallel_reps,
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,