+++ R/tar_map2_raw.R
@@ -11,7 +11,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
-#' @inheritSection tar_rep Nested futures for batched replication
+#' @inheritParams tar_rep
 #' @param name Character of length 1, base name of the targets.
 #' @param command1 Language object to create named arguments to `command2`.
 #'   Must return a data frame with one row per call to `command2`.
@@ -48,6 +48,7 @@ tar_map2_raw <- function(
   columns2 = quote(tidyselect::everything()),
   suffix1 = "1",
   suffix2 = "2",
+  parallel_reps = FALSE,
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
@@ -122,7 +123,12 @@ tar_map2_raw <- function(
   )
   target_downstream <- targets::tar_target_raw(
     name = name_downstream,
-    command = tar_map2_command_downstream(command2, sym_upstream, columns2),
+    command = tar_map2_command_downstream(
+      command = command2,
+      sym_upstream = sym_upstream,
+      columns2 = columns2,
+      parallel_reps = parallel_reps
+    ),
     pattern = rlang::call2("map", sym_upstream),
     packages = packages,
     library = library,
@@ -185,12 +191,18 @@ tar_map2_command_upstream <- function(command, group) {
   )
 }
 
-tar_map2_command_downstream <- function(command, sym_upstream, columns2) {
+tar_map2_command_downstream <- function(
+  command,
+  sym_upstream,
+  columns2,
+  parallel_reps
+) {
   rlang::call2(
     "tar_map2_run",
     command = command,
     values = sym_upstream,
     columns = columns2,
+    parallel_reps = parallel_reps,
     .ns = "tarchetypes"
   )
 }
@@ -223,33 +235,71 @@ tar_map2_group <- function(data, group) {
 #'   this function directly.
 #' @return A data frame with a `tar_group` column attached (if `group`
 #'   is not `NULL`).
+#' @inheritParams tar_rep
 #' @param command Command to run.
 #' @param values Data frame of named arguments produced by `command1`
 #'   that `command2` dynamically maps over. Different from the `values`
 #'   argument of `tar_map2()`.
 #' @param columns tidyselect expression to select columns of `values`
 #'   to append to the result.
-tar_map2_run <- function(command, values, columns) {
+tar_map2_run <- function(command, values, columns, parallel_reps) {
   command <- substitute(command)
   columns <- substitute(columns)
+  columns <- targets::tar_tidyselect_eval(columns, colnames(values))
   splits <- split(values, f = seq_len(nrow(values)))
-  out <- furrr::future_map(
-    .x = seq_along(splits),
-    .f = ~tar_map2_run_rep(
-      rep = .x,
-      command = command,
-      splits = splits,
-      columns = columns,
-      reps = length(splits)
-    ),
-    .options = furrr::furrr_options(seed = TRUE)
+  reps <- length(splits)
+  call <- quote(
+    function(.x, .y, command, reps, columns) (
+      tarchetypes::tar_map2_run_rep(
+        rep = .x,
+        values = .y,
+        command = command,
+        reps = reps,
+        columns = columns
+      )
+    )
   )
+  fun <- eval(call, envir = targets::tar_option_get("envir"))
+  if (parallel_reps) {
+    out <- furrr::future_map2(
+      .x = seq_along(splits),
+      .y = splits,
+      .f = fun,
+      .options = furrr::furrr_options(
+        seed = TRUE,
+        packages = targets::tar_definition()$command$packages
+      ),
+      command = as.expression(command),
+      reps = reps,
+      columns = columns
+    )
+  } else {
+    out <- map2(
+      x = seq_along(splits),
+      y = splits,
+      f = fun,
+      command = as.expression(command),
+      reps = reps,
+      columns = columns
+    )
+  }
   do.call(vctrs::vec_rbind, out)
 }
 
-tar_map2_run_rep <- function(rep, command, splits, columns, reps) {
+#' @title Run a rep in a `tar_map2()`-powered function.
+#' @export
+#' @keywords internal
+#' @description Not a user-side function. Do not invoke directly.
+#' @return The result of running `expr`.
+#' @param rep Rep number.
+#' @param values Data frame of mapped-over values.
+#' @param command R command to run.
+#' @param reps Number of total reps.
+#' @param columns Expression for appending static columns.
+#' @examples
+#' # See the examples of tar_map2_count().
+tar_map2_run_rep <- function(rep, values, command, reps, columns) {
   envir <- targets::tar_envir()
-  values <- splits[[rep]]
   names <- names(values)
   lapply(
     X = seq_len(ncol(values)),
@@ -278,7 +328,6 @@ tar_map2_run_rep <- function(rep, command, splits, columns, reps) {
       code = eval(command, envir = targets::tar_envir())
     )
   )
-  columns <- targets::tar_tidyselect_eval(columns, colnames(values))
   out <- tar_append_static_values(out, values[, columns])
   out[["tar_batch"]] <- as.integer(batch)
   out[["tar_rep"]] <- as.integer(rep)