+++ R/tar_map2_size_raw.R
@@ -19,7 +19,7 @@
 #'   See the "Target objects" section for background.
 #' @inheritSection tar_map Target objects
 #' @inheritSection tar_rep Replicate-specific seeds
-#' @inheritSection tar_rep Nested futures for batched replication
+#' @inheritParams tar_rep
 #' @param size Positive integer of length 1,
 #'   maximum number of rows in each batch for
 #'   the downstream (`command2`) targets. Batches
@@ -64,6 +64,7 @@ tar_map2_size_raw <- function(
   suffix2 = "2",
   columns1 = quote(tidyselect::everything()),
   columns2 = quote(tidyselect::everything()),
+  parallel_reps = FALSE,
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
@@ -98,6 +99,7 @@ tar_map2_size_raw <- function(
     suffix2 = suffix2,
     columns1 = columns1,
     columns2 = columns2,
+    parallel_reps = parallel_reps,
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,