+++ R/tar_package.R
@@ -10,7 +10,7 @@
 #' @importFrom dplyr bind_rows mutate select
 #' @importFrom fs dir_create is_dir path_ext path_ext_remove
 #'   path_ext_set path_rel
-#' @importFrom furrr future_map furrr_options
+#' @importFrom furrr future_map future_map2 furrr_options
 #' @importFrom rlang as_function call2 check_installed enquo expr
 #'   inform is_missing quo_squash
 #' @importFrom targets tar_assert_chr tar_assert_dbl tar_assert_df