+++ tests/testthat/test-tar_quarto_rep.R
@@ -173,7 +173,8 @@ targets::tar_test("tar_quarto_rep() run", {
           par = c("parval1", "parval2", "parval3-modified", "parval4"),
           stringsAsFactors = FALSE
         ),
-        batches = 2
+        batches = 2,
+        parallel_reps = FALSE
       )
     )
   })
@@ -214,7 +215,8 @@ targets::tar_test("tar_quarto_rep() run with output_file specified", {
           par = c("parval1", "parval2", "parval3", "parval4"),
           output_file = c("f1.html", "f2.html", "f3.html", "f4.html"),
           stringsAsFactors = FALSE
-        )
+        ),
+        parallel_reps = TRUE
       )
     )
   })